Separate remove_duplicates functionality into file remove_duplicates.py
